Refactor tests with common code
