+++ samples/snippets/src/test/java/com/example/bigtable/WritesTest.java
@@ -16,70 +16,28 @@
 
 package com.example.bigtable;
 
-import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThat;
 
-import com.google.cloud.bigtable.admin.v2.BigtableTableAdminClient;
-import com.google.cloud.bigtable.admin.v2.models.CreateTableRequest;
-import java.io.ByteArrayOutputStream;
-import java.io.PrintStream;
-import java.util.UUID;
+import java.io.IOException;
 import org.hamcrest.CoreMatchers;
 import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
 
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class WritesTest {
-
-  private static final String INSTANCE_ENV = "BIGTABLE_TESTING_INSTANCE";
-  private static final String TABLE_ID =
-      "mobile-time-series-" + UUID.randomUUID().toString().substring(0, 20);
-  private static final String COLUMN_FAMILY_NAME = "stats_summary";
-
-  private static String projectId;
-  private static String instanceId;
-  private ByteArrayOutputStream bout;
-
-  private static String requireEnv(String varName) {
-    String value = System.getenv(varName);
-    assertNotNull(
-        String.format("Environment variable '%s' is required to perform these tests.", varName),
-        value);
-    return value;
-  }
+public class WritesTest extends CommonTest {
 
   @BeforeClass
-  public static void beforeClass() {
-    projectId = requireEnv("GOOGLE_CLOUD_PROJECT");
-    instanceId = requireEnv(INSTANCE_ENV);
-    try (BigtableTableAdminClient adminClient =
-        BigtableTableAdminClient.create(projectId, instanceId)) {
-      CreateTableRequest createTableRequest =
-          CreateTableRequest.of(TABLE_ID).addFamily(COLUMN_FAMILY_NAME);
-      adminClient.createTable(createTableRequest);
-    } catch (Exception e) {
-      System.out.println("Error during beforeClass: \n" + e.toString());
-    }
-  }
-
-  @Before
-  public void setupStream() {
-    bout = new ByteArrayOutputStream();
-    System.setOut(new PrintStream(bout));
+  public static void beforeClass() throws IOException {
+    initializeVariables();
+    createTable();
   }
 
   @AfterClass
-  public static void afterClass() {
-    try (BigtableTableAdminClient adminClient =
-        BigtableTableAdminClient.create(projectId, instanceId)) {
-      adminClient.deleteTable(TABLE_ID);
-    } catch (Exception e) {
-      System.out.println("Error during afterClass: \n" + e.toString());
-    }
+  public static void afterClass() throws IOException {
+    cleanupTable();
   }
 
   @Test
